feat: Refactor TreeNode conversion methods and add tests
The code refactors the way TreeNodes are converted to arrays and vice versa in Golang. It removes old commented Python code and introduces new, tested Golang functions for converting an array to a TreeNode, an array to a TreeNode with a target, and a TreeNode back to an array. Also, new test cases are added to ensure the correctness of these functions.