feat: Add support for multiple programming languages in problem solutions
The scripts have been updated to support fetching of problem solutions in multiple languages. The "get_problem" script includes a new parameter "languages" and now is able to process and write solutions for both Python and Golang. Also, "daily_auto" and "daily_submission" scripts will now default to Python3 in case of an invalid or missing LANGUAGES environment variable.